requirements.txt
accelerate==1.13.0
annotated-doc==0.0.4
anyio==4.13.0
attrs==26.1.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
contourpy==1.3.3
cryptography==47.0.0
cycler==0.12.1
et_xmlfile==2.0.0
exceptiongroup==1.3.1
filelock==3.29.0
fonttools==4.62.1
fsspec==2026.3.0
google-auth==2.49.2
google-auth-oauthlib==1.3.1
gspread==6.2.1
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.12.0
idna==3.13
Jinja2==3.1.6
kiwisolver==1.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.9
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
oauthlib==3.3.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==26.2
pillow==12.2.0
psutil==7.2.2
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.20.0
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.4.4
reportlab==4.4.10
requests==2.33.1
requests-oauthlib==2.0.0
rich==15.0.0
safetensors==0.7.0
selenium==4.43.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.14.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.6.2
trio==0.33.0
trio-websocket==0.12.2
typer==0.24.2
typing_extensions==4.15.0
urllib3==2.6.3
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.3.2

# --- web stack (добавлено для FastAPI / SPA-хостинга) ---
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
psycopg2-binary>=2.9
pydantic>=2.5
pydantic-settings>=2.0
python-multipart>=0.0.9
aiofiles>=23.0
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
